chore: refactor all spawns to areas\spawns\
