+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/areas/spawns/spawns_11929.plugin.kts
@@ -1,4 +1,4 @@
-package gg.rsmod.plugins.content.areas.falador.dwarven_mines
+package gg.rsmod.plugins.content.areas.spawns
 
 spawn_item(item = Items.TOMATO, amount = 1, x = 3085, z = 3261, respawnCycles = 170)
 spawn_item(item = Items.CHEESE, amount = 1, x = 3083, z = 3260, respawnCycles = 170)
