+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/areas/spawns/spawns_12337.plugin.kts
@@ -1,4 +1,4 @@
-package gg.rsmod.plugins.content.areas.draynor
+package gg.rsmod.plugins.content.areas.spawns
 
 /**
  * @author Alycia
