+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/areas/spawns/spawns_13881.plugin.kts
@@ -1,3 +1,3 @@
-package gg.rsmod.plugins.content.areas.daemonheim
+package gg.rsmod.plugins.content.areas.spawns
 
 spawn_npc(npc = Npcs.FREMENNIK_SHIPMASTER_9708, x = 3513, z = 3694, direction = Direction.WEST, static = true)
